third_party/WebKit/LayoutTests/external/wpt/2dcontext/scroll/2d.scrollPathIntoView.verticalRL-expected.txt

This is a testharness.js-based test.
FAIL scrollPathIntoView() works in vertical-rl writing mode assert_equals: Math.round(rect.right) === viewportWidth + (canvasWidth - 4 - 16) (got 600[number], expected 865[number]) expected 865 but got 600
Harness: the test ran to completion.
